refact : add system and core folder and move files - add io class to handle system input output
