+++ src/main/java/lotto/Application.java
@@ -1,13 +1,13 @@
 package lotto;
 
-import lotto.controller.Shop;
-import lotto.view.OutputView;
+import lotto.system.controller.ShopController;
+import lotto.system.view.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
         try {
-            Shop shop = Shop.visitShop();
+            ShopController shop = ShopController.visitShop();
             shop.buyLottos();
             shop.checkLottoResult();
         } catch (RuntimeException e) {
