+++ src/main/java/lotto/core/validate/MoneyValidate.java
@@ -1,4 +1,4 @@
-package lotto.validate;
+package lotto.core.validate;
 
 public class MoneyValidate {
