+++ src/main/java/lotto/system/controller/ExceptionController.java
@@ -1,0 +1,4 @@
+package lotto.system.controller;
+
+public class ExceptionController {
+}
