+++ src/main/java/lotto/system/view/InputView.java
@@ -1,23 +1,21 @@
-package lotto.view;
-
-import camp.nextstep.edu.missionutils.Console;
+package lotto.system.view;
 
 public class InputView {
 
     public String  showInputMoneyMessage() {
         System.out.println("구입금액을 입력해 주세요.");
-        return Console.readLine();
+        return SystemIO.readUserInput();
     }
 
     public String showInputLottoNumberMessage() {
         System.out.println("");
         System.out.println("당첨 번호를 입력해 주세요.");
-        return Console.readLine();
+        return SystemIO.readUserInput();
     }
 
     public String showInputBonusNumberMessage() {
         System.out.println("");
         System.out.println("보너스 번호를 입력해 주세요.");
-        return Console.readLine();
+        return SystemIO.readUserInput();
     }
 }
