+++ src/main/java/lotto/system/view/SystemIO.java
@@ -1,0 +1,13 @@
+package lotto.system.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class SystemIO {
+
+    public static String readUserInput() {
+        return Console.readLine();
+    }
+    public static void showMessageToConsole(Object message) {
+        System.out.println(message);
+    }
+}
